requirements.txt
pymongo
newsapi-python
beautifulsoup4
urllib3
pycrypto
